Fix ConfigImpl bug and add more tests
Fix bug in ConfigImpl wherein a boolean would be converted into
a ConfigBoolean with value true even if the boolean is false.
Increase test coverage for ConfigValueFactory tests by adding
a test to ensure this bug is no longer happening, and increase
test coverage of SimpleConfig spec tests by ensuring that
data structures can be added to a config.