Fix unecessarily strict test case.
As it turns out, hocon does not actually require that the output be
rendered in the same order by every implementation so when running with
ruby-1.9.x vs ruby-1.8.7-p352 for instance the rendered string may not have
variables specified in the same order.

However, hocon does require that comments be matched to their variables. This
patch validates that behavior by creating a hash of config-lines mapped to lists
of preceding comments and verifies that this hash is the same before and after
rendering regardless of which hocon implementation created the "original" output
file.

Also, I find the input vs output semantics and the way variables with these
names are used just a little confusing but whatever.

Signed-off-by: Wayne <wayne@puppetlabs.com>