(TK-128) fix tokenization of unquoted strings
This commit fixes the tokenization of unquoted strings which might be
numbers.  In particular, this affects IP addresses.

The tokenizer aggressively attempts to parse such strings as
either a Float or Integer and relies on an error being raised to
indicate when the string is not actually a valid number.  However,
`to_i` and `to_f` never raise execptions, so the constructors must be
used instead.