(TK-246) Fix single-line config bug
Previously there was a bug wherein building out a config starting
from an empty ConfigDocument would cause the entire config to
exist on a single line. Fix this bug by modifying the addition of
new maps along a path to add multi-line maps instead of
single-line maps if the object being added to is an empty root or
a multi-line object.