(MAINT) Fix Hocon.load substitution issue
Previously, Hocon.load was calling into the
ConfigFactory.parse_file method. However, the `parse` methods
in ConfigFactory do not resolve substitutions, as that is the
intent of the `load` methods.

This commit updates the Hocon.load method to call into
ConfigFactory.load_file. It also updates the readme to explain
the usage of ConfigDocuments, and adds a comment to explain
that the `load` methods in ConfigFactory should be used if
substitutions are present