(HC-79) support `:syntax` option in simple `load`
This commit adds support for an optional `opts` map to be passed
in to the simple `load` method.  If provided, this map may contain
a `:syntax` key that explicitly specifies which config format/syntax
the user expects the file to be in.

This provides a way for users to load files whose file extension
doesn't match the built-in expectations for which file extensions
use which syntaxes.

The commit also provides some error checking for the case where
an explicit `:syntax` is not passed in, and the file extension
isn't recognized.  In this case, we will throw an error now, rather
than silently returning an empty map like we did in the past.

Finally, this commit adds some notes to the docs/example usage,
indicating how to pass in an explicit syntax.