(HC-80) Don't shadow ruby Class/Module#name method
Prior to this commit, there were a few places in the code where
we'd ported over a class-or-module-level method named `name` from
the upstream library.  This isn't a good idea in Ruby because it
results in shadowing of the built in Ruby methods Class#name and
Module#name.  This was causing problems for some users, e.g. in
cases where reflection is being used to examine classes.

In this commit we rename all such methods to something more specific,
and replace the calls to the old names with calls to the new names.