Fix bug in `Hocon.parse` with substitutions
Currently, if you call `Hocon.parse` with a string that contains
substitutions, you will get a `ConfigNotResolvedError`.

We had this issue with `Hocon.load` earlier, and modified it to
include the code necessary to resolve the config object before
returning it.  However, we didn't make the same changes for `parse`,
so the behavior actually diverged between the two.

This commit fixes up `parse` in the same way that we previously
fixed up `load`.